requirements.txt
telethon>=1.34.0
cryptography>=41.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0
colorama>=0.4.6
asyncio>=3.4.3